Create vertex-to-cell interp routine, interp. dissipation
This commit interpolates the heat dissipation term from Albany on
vertices to cell centers where the thermal code requires it.  I wrote a
wrapper to the MPAS framework Wachspress interpolation routines that
does this interpolation.  These routines could be moved to framework.

Visualization indicates that the interpolation appears to be working
correctly.