Fix units on beta, calculation of basalFrictionFlux
Ever since creating the Albany interface, we have inadvertently been
using beta in units: Pa yr/m instead of SI units.  We realized this at
some point, but changing it to SI would complicate setting up the MISMIP
tests, so we've left it for now.

This commit formalizes this convention in the metadata in Registry, and
corrects the units of beta as applied to the basalFrictionFlux
calculation.