Add accumulation fields for land-ice fluxes
These fields, accumulatedLandIceMass, accuulatedLandIceHeat,
and accumulatedLandIceFrazilMass, contain the heat and
freshwater per unit area stored in land ice due to surface
fluxes, and mass due to frazil formation under land ice.
These fields are needed to keep track of the heat and freshwater
budgets when not coupled to a land-ice model.

Add landIceFreshwaterFlux, accumulatedLandIceMass,
accumulatedLandIceHeat, and accumulatedLandIceFrazilMass, to
the globalStats AM.

Most importantly, the sum of the accumulation fields,
accumulated*Sum, are the total mass and heat stored in land
ice formed during the simulation and the mass of frazil formed
under land ice.  These fields are needed for the global freshwater
budget.