Fix three bugs in globalStats
MinVertSum and MaxVertSum:
Previously the variables verticalSumMins_tmp and verticalSumMaxes_tmp
were being used to compute min/max of vertical sums even though
they were not being initialized for computations involving
surface variables.  Now, verticalSumMins and verticalSumMaxes
are used instead

tracer1:
Stats were not being computed for tracer1 even though it was in
the registry.  This led to an indexing probelm where each stat
starting with tracer1 had the values intended for the next stat.

frazilLayerThicknessTendency:
The MinVertSum and MaxVertSum were not being given a value when
frazil is not present, so a garbage value is written out.