Add support for writing time of day in seconds
This commit adds support to stream definitions for the $S expansion
character, which expands to the number of seconds in the current day.

It can be used as a replacement for $h.$m.$s, and some frameworks
require streams to be written using a format more like $Y-$M-$D_$S to
automatically handle the output files correctly.