Subtract landIceDraft from pressureAdjustedSSH
Add a variable landIceDraft that is the initial SSH in equilibrium
with the landIcePressure.  The landIceDraft is subtracted from ssh
as part of computing pressureAdjustedSSH.  This way, the sea
surface tilt used in the sea-ice model does not include tilt due
to the depression of the sea surface under land ice.

The landIceDraft is stored in init mode as part of the balance
calculation in the ssh and landIcePressure init module.  It
is modified along with the ssh during iterative adjustment and
is read in/written out in every stream that has landIcePressure.

Also, remove unneeded interpolation in landIcePressure init.
When config_iterative_init_variable == 'landIcePressure_from_top_density',
the vertical grid is computed with the known ssh from the beginning and
there is no need to re-interpolate tracers later on.  This has been cleaned
up and simplified.

A few other bugs have been fixed:
  * when config_iterative_init_variable == 'ssh' or 'landIcePressure',
    the ssh and landIcePressure are set to zero when modifySSHMask == 0
  * activeTracers beyond maxLevelCell are now given a fill value rather
    than having a garbage value