Enable the init core to report minimum necessary value for config_nfglevels
When the number of first-guess levels exceeds the value specified for
the namelist option config_nfglevels, the code now reports the minimum value
that this namelist option must be set to. Previously, the code would simply
say that config_nfglevels must be increased, but users without access to
tools like the WPS rd_intermediate utility may not have an easy way of knowing
how many levels they had in their intermediate files.