Add script to make config files for param. study
The script uses a template file that contains dummy replacement strings to
insert the values of parameters.  The resulting config files are numbered
consecutively with a prefix povided.  Parameter names and values are provided
as a list as in this example:
-p param1=1,2,3 param2=1e3,1e4,1e5 param3='a','b','c' \
   param4=.true.,.false.,.true.
The number of parameter values must be the same for all parameters and all
parameters are varied simultaneously.