Add script to setup restarts
The script modifies a specified namelist file.  If the
file named by config_restart_timestamp_name exists, set
config_do_restart = .true. and config_start_time = 'file'
to enable a restart run.

This can be added to the end of a job script that will be run
multiple times.  The first run is not a restart run but subsequent
runs will be.  This way, there is no need to keep multiple copies
of the namelist file (or multiple steps) that are identical except
for the two entries related to restart.