Fix print statements for L1 and L2 norms
In test_cases/ocean/utility_scripts/compare_fields.py, the pass
thresholds for L1 and L2 norms were accidentally switched in print
statements telling the user what the thresholds are.  This only affects
the value written to stdout and not the comparison itself, but it can
also lead to an error if only one of the two norms is set.