Add limiter on interface salinity in land-ice fluxes
The limiter prevents division by zero when the interface salinity
is zero and removes an error message that often halted the model.

This merge also uses landIceMask to determine where all boundary-
layer physics computations should be performed.  Previously, the
interface and boundary-layer temperature and salinity were being
computed everywhere in the domain and only the fluxes were being
masked to the ice-shelf region.