Added Experiment G in the EISMINT2 test infrastructure
I added Experiment G from Payne et al. (2000) to the EISMINT2 test suite.
Experiment G is similar to the other experiments, except that there is a specified
basal traction parameter that allows sliding everywhere on the domain.
I modified the EISMINT2 scripts accordingly, with a uniform value of beta = 1000 Pa/(m/a)
initialized in landice_grid.nc. The test appears to set up correctly.
It cannot currently be tested with the SIA dycore (which does not allow basal sliding),
but I will test it with the the HO dycore.