Replace subroutine mpas_atm_get_bdy_tend with functions for getting tendency and state
The mpas_atm_boundaries module previously had an example subroutine,
mpas_atm_get_bdy_tend, for obtaining both the tendency and the state for boundary
variables. This commit replaces this subroutine with two separate functions,
mpas_atm_get_bdy_tend and mpas_atm_get_bdy_state, which return as an array
the tendency for the specified boundary variable or the state of the specified
boundary variable.