Retain private module pointers to 'clock' and 'blocklist' in atm_time_integration module
At the beginning of the atm_timestep routine, we set pointers to domain % clock and
domain % blocklist so that we don't need to pass domain, clock, etc. to any subroutine
that needs these in, e.g., calls to mpas_atm_get_bdy_state.

Note that routines in atm_time_integration not called below the call to atm_timestep
in the call tree will not have access to these pointers until the first call to
atm_timestep. This can affect routines such as atm_compute_solve_diagnostics.