import swapm implementation of MPI_AlltoallV from PIO
For high resolution runs of ACME and CESM using large
MPI process counts, the existing MPI algorithms in the
rearrange_ routine can be inefficient. The swapm variant
of the MPI_AlltoallV operator was ported from PIO1 and
modified to work in the MCT environment. The option to
call this was then added to rearrange_ . For the short
term, rearrange_ only calls swapm (and not MPI_AlltoallV
for the existing MPI-1 point-to-point communication
algorithm) and uses fixed swapm parameter options that
are anticiapted to be resonable choices for most
situations. Long term, the routines calling rearrange_
should be modified to allow the user to specify an MPI
algorithm and communication protocol.