Obviate the need for, and revert, the TEMPORARY FIX in eea47d2
To work around the issue of variables with packages being passed through
argument lists, we can simply check if the pointer to the variable is
valid, and if not, allocate it with zero size before the function call.

With the changes in this commit, we can now reinstate the use of physics
packages for variables.