restrict swapm option to calls from sMatAvMult_SMPlus_
Based on performance experiments on Titan, using the
swapm option with handshaking enabled for all calls to
rearrange_ is not as efficient as using this option only
in calls to rearrange_ from sMatAvMult_SMPlus_ . Changed
logic to implement this restriction. While it may be
useful in other locations as well, these locations
are still to be determined.