Pre-multiply 'zb3_cell' and 'adv_coefs_3rd' by 'config_coef_3rd_order' at start of run
Everywhere zb3_cell and adv_coefs_3rd are used in the solver, they are coupled to
config_coef_3rd_order. Because these fields are constant throughout a run, we can
perform the multiplications once at the start of the simulation with a new routine
atm_couple_coef_3rd_order() that is called from atm_mpas_init_block().

Note that atm_couple_coef_3rd_order must be called before atm_init_coupled_diagnostics,
but after the call to atm_adv_coef_compression; this necessitated the move of
the call to atm_adv_coef_compression to the top of the atm_mpas_init_block routine.