Add new mpas_atm_diagnostics_utils module for diagnostics framework
This new module is intended to provide utility routines that can be used
by diagnostics routines. The initial implementation of the module includes
a function mpas_field_will_be_written that can be used to determine whether
a field will be written by the next call to mpas_stream_mgr_write.

The initial implementation of the mpas_atm_diagnostics_utils module requires
a reference to the stream manager, and this is now passed through the call
to mpas_atm_diag_setup, which in turn calls mpas_atm_diag_utils_init to
save the necessary stream manager reference.

This commit also adds calls to mpas_atm_diag_compute and mpas_atm_diag_reset
before and after the initial streams are written (before the first timestep).