made transect configurable in transect transport AM changed transect transport am so that it uses a single transect group
changed the dimension for the outputs of the MOC streamfunction to be 3
made it so that the AM wont crash when a region group with more than 3 regions in it is configured for this AM
the AM will still crash horribly when the number of transects in the selected transect group is not the same as the number of regions in the selected regiongroup.

changed the maximum number of MOC regions down to R3 (three) from max_regions_in_group, becuase there will be no more than 3 MOC regions plus the global MOC, which is separate anyways.

ran the linter over the code and fixed every error it showed