Bugfix for failing C compset DEBUG tests
When water isotopes were added to the coupler, the runoff -> ocean fields
changed -- this is fine in coupled and ocean / ice cases, but the new isotope
fields should not be used in ocean-only compsets. To get around this, I
introduced two new module variables (seq_flds_r2o_liq_fluxes and _ice_fluxes)
in seq_flds_mod.F90 that always contain 'Forr_rofl' and 'Forr_rofi',
respectively, and can also include the isotope fluxes if flds_wiso is .true.

The call to seq_map_map uses these new variables as the fldlist arguments in
calls from prep_ocn_calc_r2x_ox(), so that we don't run into the issue of
trying to map non-existent fields when water isotopes are not enabled.