replacing peLocs_ algorithm with high performance alternatives
The complexity of the current peLocs_ algorithm is O(npoints*ngseg),
which has been observed to take over 10 minutes for certain high
resolution runs of ACME and CESM. Two alternative algorithms
are added to the existing one, controlled by ifdefs. The new default
algorithm has a worst case complexity of O(gsize + ngseg), and in
practice is typically much smaller than this. In the same example problem,
the new default is 60 times faster than the original algorithm.
The downside is that the worst case memory requirement is O(gsize).
This additional space is allocated and deallocated within the routine
and, in practice, the memory requirements have been observed to be
much smaller, but this is not guaranteed. Another alternative algorithm
was also implemented, and is enabled if compiled when the CPP flag
MEDIUM_MEMORY_PELOCS is defined. The performance of this version
has been observed to be within 20% of that of the default,
and the worst case memory requirements are O(ngseg). The original
algorithm is enabled by compiling with the CPP flag LOW_MEMORY_PELOCS
defined. Note that LOW_MEMORY_PELOCS takes precedence over
MEDIUM_MEMORY_PELOCS.