Diff. freezing temp. below land ice and in open ocn
This merge changes the freezing temperature to have a different set
of coefficients in the open ocean an in land-ice cavities.  The
default coefficients are changed so the freezing temperature is
constant (-1.8 C) in the open ocean, and is fit to the TEOS-10
potential freezing temperature in land-ice cavities using
20 <= S < 40 PSU
0 <= p <= 2e7 Pa (~ 2000 m)

All calls to ocn_freezing_temperature have been updated to indicate
whether the freezing temperature is below land ice or in the open
ocean (possibly below sea ice).