Enable multiline abort messages from mpas_dmpar_global_abort
This commit moves the mpas_dmpar_global_abort routine to a new module, mpas_abort,
and adds a second, optional argument to the routine after the message argument. This
new argument is a logical that tells the mpas_dmpar_global_abort routine whether to
actually call MPI_Abort or to simply append the message argument to a fatal error
file.

By placing the mpas_dmpar_global_abort routine in its own module, we avoid any
problems with circular module dependencies, and we allow the compiler to check for
the optional argument, deferredAbort, to the routine.

All modules that made calls to mpas_dmpar_global_abort have been modified to include
the line "use mpas_abort, only : mpas_dmpar_global_abort".

Additionally, the name of the fatal error file has been changed with this
commit to the form 'log.????.abort', where ???? is expanded up to 9 digits
depending on the size of MPI_COMM_WORLD.