Add 'core_atmosphere/utils' directory with dummy program to build physics tables
This commit simply introduces changes to the MPAS-Atmosphere directory and build
structure to allow us to add extra utility programs to be compiled along with
the 'atmosphere_model' executable at run-time. An example of where such a utility
may be needed is in the pre-computation of large physics lookup tables deemed too
large to download at build time.