Optimize isobaric diagnostics by only computing fields if they will be written
This commit adds logic to only compute isobaric diagnostic fields if they will
be written in the timestep in which isobaric_diagnostics_compute( ) is called.

Some inefficiency is still present, since, e.g., if any pressure level of relative
humidity will be written, all available levels of RH are interpolated. At present,
the extra expense (mostly in terms of code logic) to handle each level individually
did not seem to be worth the savings in computation.