Reset alarm on 'iau' stream in atm_core_init( )
After the logic to read the 'iau' stream when IAU is active, we should always
reset the alarm for the stream regardless of IAU to ensure we don't read
the stream in the mpas_stream_mgr_read( ) calls in atm_core_run( ).

This commit also removes an unused variable 'tend_iau'.