Rename the IAU module to "mpas_atm_iau"
Because the module computes weights in addition to applying analysis increments, and
since the functionality in the module may expand further with future development,
the module has been renamed to the more general name "mpas_atm_iau".

This commit also adds a copyright statement to the top of the mpas_atm_iau module,
and removes an unused use of the old module in the mpas_atm_core module.