Move pool accesses into IAU routines to tidy up atm_srk3 code
This commit moves acccesses of configuration options and subpools from
the main atm_srk3 routine to the beginning of the IAU routines that use
those pool members.

This commit also changes the atm_compute_iau_coef subroutine into
a function named atm_iau_coef, since the objective of the routine is
to produce a single value as output. This could in principle allow uses
of the coefficient produced by the function to use an inline function call.