Simplify and make robust the handling of scalars in IAU module
The IAU module previously handled all analysis increments for scalar
constituents individually, which could lead to problems when some scalars
are not allocated due to packages.

To simplify matters, the "amb" scalars have been moved into a var_array
just like the regular scalars and scalars_tend var_arrays; and, each
constituent in the new "scalars_amb" var_array the same packages attached
to it as its counterpart in the regular scalars array.

Now, all scalars can be handled in the IAU module with the use
of simply scalars(moist_start:moist_end,:,:).

Note: We should consider how to handle cases where one of the three scalars
var_arrays (scalars, scalars_tend, scalars_amb) in the Registry is updated
in a way that is inconsistent with the others.