Add code for printing the location of the min/max winds in the global domain
This new code follows the functionality already available in the model to print
the min/max values in the u and w field at the end of each timestep, but adds
information on the model level, latitude, and longitude where those extrema occur.

Also included in the new code are checks for NaNs in the u and w fields. If any
NaNs are detected, the model is stopped.

This printout is enabled by setting a new namelist option
'config_print_detailed_minmax_vel' to .true.. The resulting printout looks like
the example below.

 global min w:   -0.6845 k=  47,  -0.957 lat  -69.221 lon
 global max w:    0.8648 k=  36,  42.058 lat  -92.353 lon
 global min u: -117.5393 k=  55, -55.012 lat   23.458 lon
 global max u:  117.2959 k=  55, -55.235 lat   25.359 lon
 global max wsp:  117.9053 k=  55, -55.012 lat   23.458 lon