Bug fix for soundings module when config_sounding_interval = 'none'
Previously, the simulationClock pointer was not set until after
the 'sounding_locations.txt' file was read, which did not happen
if config_sounding_interval was set to 'none'. This could lead to
simulationClock being a null pointer, which would cause a segfault
in the call to soundings_compute( ).

Now, we set simulationClock before potentially returning from
soundings_setup( ) when config_sounding_interval = 'none' or when
there is no sounding_locations.txt file.