Calculate GWD fields on native MPAS mesh when config_native_gwd_static=true
This commit introduces a new namelist option, config_native_gwd_static, which,
when set to true, invokes code in the mpas_init_atm_gwd module to compute
sub-grid-scale orography statistics directly on the native MPAS mesh.

Also included in this commit is the introduction of a new package, gwd_stage_in,
which is used to ensure that static fields will be read if we will be computing
GWD static fields but will not first interpolate other static fields.