Add 'config_gwd_cell_scaling' option to control area over which GWD fields are computed
In some experiments with the GWDO scheme in MPAS-Atmosphere, we have found it beneficial
to compute sub-grid-scale orographic statistics over an area that is larger than
the cell to which the statistics belong. This commit adds a new namelist option,
config_gwd_cell_scaling, that defaults to 1.0 and is not in the namelist.init_atmosphere
file by default; this coefficient is multiplied by the mean cell diameter to determine
the width and height of the box over which GWD statistics are computed.