Delete unused GPTL macros from dynamics code
There were several GPTL macros that could be enabled if the MPAS code
were linked with the GTPL library. However, these were only used while
optimizing the dynamics, and rely on extra code that is not present in
the top-level Makefile. For now, we just remove these macros to clean
up the dynamics code.