Add a check in init_case 8 to verify that we have a valid 'landmask' field
If the global maximum of the 'landmask' field is zero, then we either have
some sort of aquaplanet, or the 'landmask' field was not in the input file.
For now, it seems best to assume that init_case 8 will not be used for
aquaplanet-type simulations, and to stop if no land points were found when
interpolating surface fields.