Prevent 'build_tables' from being run in parallel
The 'build_tables' utility is currently intended to be a serial-only program,
but since MPAS-Atmosphere itself is parallel, some people may mistakenly
run 'build_tables' in parallel. To avoid this situation -- which could allow
multiple copies of the executable to write to the same file -- we add
"status='new'" and "iostat=istatus" to the open( ) statements for the tables.

At least in C, the open( ) statement guarantees the following:

  O_EXCL Ensure that this call creates the file: if this flag is specified in
         conjunction with O_CREAT, and pathname already exists, then open()
         will fail. The behavior of O_EXCL is undefined if O_CREAT is not specified.

Under the assumption that "status='new'" enforces similar behavior in Fortran,
we can guess if (istatus /= 0) that the table file already exists from a previous
execution of the utility, or that there may be multiple instances of the utility
running in parallel and the current instance is not the one who was able to
uniquely create the table file. Then, we can print an error message and quit.