decrease serial complexity of initp_ algorithm
These changes implement two different optimizations
for the initp_ routine in m_Router.F90

The first optimization uses a preprocessing step to
attempt to determine a subset of processes to loop
over, so as to decrease the serial complexity of a
later, expensive, nested loop. It also uses this to
decrease the size of the local arrays
allocated/deallocated within the routine. In current
practice in ACME, this does very little, as the
preprocessing step seldom decreases the number of
iterations of the expensive nested loop much. As
there is little cost to this, leaving it in in case
it is useful in the future.

The second optimization adds additional logic to the
expensive nested loop to eliminate iterations of the
inner loop that are known to be unnecesary based on
preprocessed information and from information gained
from earlier iterations of the outer loop. This decreases
the cost of this nested loop significantly. This is
of little importance on Titan, decreasing intialization
cost by a minute or two, but it decreases the cost of
intialization on Mira from 106 minutes to 26 minutes
for a production-like high resolution coupled model runs.
While affecting only the initialization, the high
initialization cost on Mira makes development and testing
very difficult.