Initialize garbage elements for locally allocated arrays in dynamics
There are a number of arrays that are local to the dynamics that we
allocate and deallocate in each dynamics step (e.g., s_max_arr, s_min_arr,
delsq_u, etc.). These arrays were previously not initialized, and this
lead to situations where the otherwise inconsequential use of the garbage
element caused a floating point exception.

To avoid the generation of FPEs in the dynamics, we now initialize
the garbage element of all locally allocated arrays to 0.0_RKIND.