disable MPI_RSEND by default, but retain option to re-enable
MPI_RSEND and MPI_IRSEND are not implemented correctly in
recent versions of the MPI library provided by Cray. This
modification maps these to MPI_SEND and MPI_ISEND, respectively,
except when the CPP token _USE_MPI_RSEND is defined during
compilation. Note that the default is now NOT to use MPI_RSEND.
The use of MPI_RSEND and MPI_IRSEND should be a performance
enhancement, and has been in the past and with other compilers,
so prefer to retain this as an option for future evaluation.