Robustness tweaks to setup MISMIP P75S script
This makes some minor changes to improve the usefulness and robustness
of the script to set up the P75S experiment.  Note there still is some
unresolved strange behavior, I think when the restart file as multiple
time levels in it, but a workaround is to use ncks to pull out the final
time level and/or run the script once with the restart file included and
once without it.