Fix perturb script to check number of time levels
Previous erratic behavior from this script occurred when the output and
restart files being queried had only one or more than one time level
(I'm not sure which).  This commit adds logic to determine the number of
levels in each and access fields accordingly.  This seems to fix the
mysterious failures seen previously, but I have not done exhaustive
testing.