Remove duplicate files and update cmake list for PRIM
Remove files from the prim directory that were duplicated in share directory. Two of the
files, prim_driver_mod.F90 and prim_advance_mod.F90, are shared across different build
targets and belong in the share directory. prim_implicit_mod.F90 is specific to implicit
time stepping and could be moved back to the prim directory along with other modules in
the share directory that are not used by other targets.