Fixes a bug where snicar file path was hardwired in ice_init.F90
In ice_init.F90 file of cice compnenet, snicar file path was hardwired.
This component is used only for F compset cases. This PR fixes that so
that the file name is now read through ice_in namelist. Checks are also
added so that the code exits if the file or any of its variables are
not present.

[Non-BFB] - Non Bit-For-Bit