Replace strncpy and snprintf where applied to non-null-term. strings
Valgrind is complaining when strncpy or snprintf are used with
non-null-terminated strings. Apparently some versions of these
routines call strlen on the input string as an sanity check on the
passed-in length parameter, and this is leading to the access of
random memory. (Note that this perhaps is a performance issue, but
it has not affected correct functioning - this code has been in
there for a long while.)

To eliminate the issue, strncpy and snprintf are replaced with
equivalent character-by-character copy loops in the
instances in which the input string is non-null-terminated.