Update ChangeLog to correct original description of the problem
The original description of the source of the problem indicated
that it was the use of strlen in strncpy and snprintf when
copying non-null-terminated strings. This would be a problem,
but the command actually used was strnlen, which should have worked
fine (and has until valgrid flagged this as a problem). Modified
description to correct name of routine, and to indicate that the
actual reason for the failure is not understood.