Fix for 'cape', 'cin', 'relhum', and isobaric 'dewpoint' computation
The 'relhum' field, which is used in the computation of the CAPE, CIN, and
isobaric dewpoint diagnostics, is given as a percent, while the computation of
the dewpoint field used in these diagnostics assumed that 'relhum' was given as
a fraction. The dewpoint computation in both the isobaric_diagnostics and
convective_diagnostics modules applied an upper bound of 1.0 on the 'relhum'
field, which lead to bad values for relative humidity as well as dewpoint
temperature.

Rather than modifying the 'relhum' field by placing an upper bound, we now
compute a temporary, fractional relative humidity for use specifically in
the computation of the 'dewpoint' field, which is in turn used by the routines
for computing CAPE and CIN.