Add halo exchange for 'exner' to recover bit-identical restartability with physics
In order to obtain bit-identical restarts when running with physics, add a halo
exchange for the 'exner' field at the beginning of each dynamics subcycle in
the atm_srk3( ) routine. Alternatively, we could have added two halo exchanges,
one at the end of the first (n-1) dynamics subcycles and one after the call to
driver_microphysics( ).