Add print statements to identify default real precision in log files
This commit adds print statements in the atm_setup_core() and init_atm_setup_core()
routines to identify the default real precision that was used when the core was
built. These routines are called early on in the model startup sequence, and will
result one of the two print statements

 Using default single-precision reals

or

 Using default double-precision reals

showing up at the top of the log.*.err files.