Compute dewpoint in isobaric and convective diagnostics following Bolton
For consistency with the computation of dewpoint temperature in DART, and to avoid
complications of the computation of relative humidity influencing the computation
of the dewpoint temperature, compute 'dewpoint' in both the isobaric and convective
diagnostics modules directly from vapor pressure following
Eqn. 10 of Bolton (1980): "The computation of equivalent potential temperature".