Default PE Layout for A_WCYCL ne4_oQU240 on Skybridge and Melvin
Created three levels of PEs (S,M,L) for A_WCYCL ne4_oQU240 on skybridge and one
level for the same on Melvin. This layouts were picked for a balance of
computation cost and speed. The L is purely for speed and doesn't consider too
much the cost of running such a PE layout. The M layout is the best of both
worlds, i.e. fast yet not too expensive, while the S layout is mostly about
reducing cost while trying to still be fast. The Melvin layout is a copy of the
S layout for skybridge.

[BFB]