Initialize t_oml_200m_initial field when climatological OML depth is used
When the namelist option 'config_oml_hml0' is zero, the ocean mixed-layer depth
was initialized from a climatological field, but the top 200 m layer mean temperature,
as well as the zonal and meridional velocities, were not initialized.

Now, the initial top 200 m layer mean temperature (t_oml_200m_initial) is set to
(SST - 5) when config_oml_hml0 == 0. Also, hu_oml and hv_oml are initialized to zero.

The t_oml_200m_initial field has a small function in OML: it acts to limit the mixing
if the skin temperature is cooler than the value in this field. Without it being
properly defined, the OML model may keep cooling the ocean.