Simplify IAU calls from atm_srk3( )
This commit simplifies the interface to IAU code from the atm_srk3( ) routine.
Rather than obtaining the IAU weight for each timestep with a call to
atm_iau_coef( ) in atm_srk3( ), then passing this weight to
the atm_add_tend_anal_incr( ) routine after checking that the weight is
non-zero, we let the atm_add_tend_anal_incr( ) routine handle all details of
obtaining the weight for each timestep and returning if there is no analysis
increment to be added.