Fix bug in storage and application of IAU tendency contributions
The tendency contributions from IAU were previously stored in 'tend_rho', 'tend_u',
'tend_theta', and 'tend_scalars'. However, due to rearrangements in the solver
since the time the original IAU code was written, we now need to store the tendency
contributions to u and theta in the arrays 'tend_ru_physics' and 'tend_rtheta_physics',
which are local to the dynamics module, to ensure proper application of these
tendencies. Additionally, we need to create a new array to store the tendency
contributions from IAU to rho.