Fix incorrect assignment of 'xlamdd', 'frh', and 'mentrd_rate' in Grell-Freitas scheme
These three variables are dimensioned by (its:ite) in the MPAS version of
the GF scheme, but were incorrectly used in several places as though they were
scalars (which they are in the WRF version of the scheme). This led to an inability
to get identical results from the GF scheme for different MPI task counts.