Avoid bit-reproducibility errors in GWDO scheme due to 'kpblmin' and 'kpblmax' computation
The 'kpblmin' and 'kpblmax' values in the GWDO scheme are computed based on values
of 'kbl' for each column that is owned by an MPI rank. Because the PBL structure
is necessarily different for distinct groups of columns, the values of 'kpblmin'
and 'kpblmax' differ among MPI ranks; furthermore, the value of 'kbl' may be modified
after 'kpblmax' and 'kpblmin' are computed. To ensure identical results across
different MPI task counts, we need to recompute final values for 'kpblmin' and 'kpblmax'
after the last modification of 'kbl'.