This merge fixes a restartability error in the new Tiedtke scheme.
The new Tiedtke scheme relied on the value of 'itimestep' -- the time step
counter -- to initialize the variables 'q1b' and 't1b' to either zero or to
the values of 'qvften' and 'thften', respectively. Because the time step counter
begins at 1 even for restart runs, this caused the 'q1b' and 't1b' fields to be
set to zero in the first call after a restart, breaking bit-identical restarts
with the scheme.

Now, the dummy argument 'itimestep' in both Tiedtke schemes is passed a flag that
is set to either 0 or 1 in the convection driver depending on whether
itimestep == 1 and config_do_restart == false.