Fix issue with 'dzs' when running r8 model with r4 initial conditions
The initial conditions file provides values for 'dzs', and for the Noah
LSM, we were previously reinitializing the 'dzs' field only for land cells.
However, the Noah scheme takes as input a single column of soil layer
thicknesses, and this column is taken from the first column of 'dzs' owned
by each task.

If the initial conditions file is in single precision, and the first column
owned by an MPI task is a water column, then the Noah scheme will see soil
layer thickness values that represent 32-bit values converted to 64-bit values,
which are not the same as 64-bit values initialized directly (as we do for
land columns) for any values that are not exactly representable in floating
point (e.g., 0.1). This causes bitwise differences in runs with different
MPI task counts.

Now, we reinitialize 'dzs' for all columns, both land and water.