SHMIP viz script: Exclude N/S rows from flux calculations
The sheet flux visualized for SHMIP is calculated from the reconstructed
X velocity at cell centers.  However, the value of that field in the
north and south rows is unrealistic due to the no-flow lateral boundary
conditions.  Therefore the minimum (and mean) sheet flow values
calculated are misleading.

This commit excludes the north and south rows from the calculation of
sheet flow rate in the viz script.