SHMIP: Adjust domain setup to exclude hexagaon 'tents' on north and south
I've determined that the proper effective width of the SHMIP domain
(using the no flux north and south b.c.) is at the second to last vertex
in the north and south.  (Or the location could be described as the
outer vertex of the vertical (N/S) edge on the cells in the north and
south rows.)  This is because the extra 'tent' regions of the cells that
poke out beyond that point have no way for flux in the x-direction
because of the no-flow b.c.  Thus they consitute 'dead space'.

This commit makes two changes to account for this:
1. The magnitude of externalWaterInput is reduced (using a factor of
2/3) in cells along the north and south rows so that the dead space is
effectively receiving no source term.
2. The number of rows in the y-dir for each resolution have been
recalculated/checked to achieve the value closest to 20km width using
the above-described points on the domain to define the relevant width.

Testing shows that failing to account for 1 introduces noticeable error
into the solution (and effectively makes the solution for
non-channelizing runs become resolution-dependent as this error gets
smaller as the cell size gets smaller).

I have also added a README to the standard_resolution folder detailing
this.