SHMIP: initial creation of E tests
This just adds E1 and the basic scripts to set it up.  It has not been
tested and probably needs some tweaks yet.
E1 and E5 geometry look correct, but I have yet to run a simulation.
I am using 100m resolution for now.  This is chosen to be fine enough to
resolve glacier width.  It may be overly restrictive on the DCFL.