Introduce a new, scale-aware 3-d divergence damping scheme for acoustic modes
This commit replaces the old divergence damping mechanism with a re-engineered
filter that is scale-aware. Analysis of the linear behavior of the filter as
it was configured for the v5.0 release suggested that it should be stable for
all typical MPAS variable-mesh configurations, but certain instances were
discovered in which the filter caused the model to become unstable.
To stabilize the filter, it has been configured such that, at any place on
a variable-resolution mesh, the divergence damping term has the same effective
physical eddy viscosity as it would have for a global uniform mesh with
the same mesh spacing. Thus, the damping rate in the 3D divergence damping
filter is treated in the same scale-aware manner as the other horizontal
filters. Also, the application of the 3D divergence damping has been moved
immediately after the acoustic step to allow for the application of the filter
on the final acoustic step in the Runge-Kutta timesplit integration. Finally,
this commit changes the default value of the divergence damping filter
coefficient config_smdiv = 0.1.