SHMIP: Add test D1
This adds the setup files for creating test D1.  The other D tests will
be added once D1 has been tested.

This also adds the time-varying forcing required for D to the MPAS code.
I put it there instead of as an input because we don't yet have an easy
way to create time-varying forcing, especially if it is periodic in
time.  Also, because of the short hydro time steps, reading the forcing
from a file would result in a huge number of file reads, likely
impacting performance significantly.